Fix handling of daylight savings time.
This change fixes the service's handling of daylight savings time.

Previously, the web client would send up its numeric time zone offset
on every request, and that offset would be used to convert any local
times the user submitted to UTC (to store in the database).

However, the numeric timezone offset is not enough information to
correctly convert the time to UTC. The offset is only accurate as of
the time of the request. The correct conversion depends on the time the
user entered, the user's actual timezone (such as "PST"), and a time
zone library that knows whether daylight savings time was (or will be)
observed at the time the user entered. This change uses NodaTime to
calculate the correct conversion.

For now, the time zone will be hard-coded to pacific time. The next
logical step would be to store a classroom-specific time zone in the
database, and for the TimeZoneProvider to use that setting. That will
come in a future change.