Allow a class to be archived
This change allows a superuser to archive a class. Archiving a class
renames an existing class, and creates a new class with the same name
as the existing name. The new class includes all projects, questions,
and assignments from the original class, but no sections or per-student
data (such as commits or submissions). The old class is marked inactive
(prohibiting non-admins from accessing it).