Add navigation options and progress information to the question solver
This change adds a navigation interface to the question solving UI.
Students can now easily navigate to the next or previous question in
the assignment. In addition, students can see the completion status of
every question in the assignment directly from the question solver, and
jump directly to any question in the assignment. This replaces the old
question solving workflow, in which students opened each question in a
separate tab from the assignment results page.