Add the ability to post announcements
This change adds the ability to post announcements to one or more
sections of a class. Teachers author announcements with a rich text
editor, and select the sections the announcements should go to. The
posted announcement appears on the home page for the class. It is also
e-mailed to all students in the selected sections, and all admins in
the class.

After posting an announcement, an announcement can be edited. Edits do
not get re-sent to students that already received the announcement, but
do get sent to any students in newly-selected sections (as the students
in those sections have not yet received the announcement). In addition,
announcements can be deleted (causing them to no longer appear on the
home page of the class).