Hide assignments not yet due from updated scores view
When the updated scores view shows a list of all scores that have
changed since a given date, it currently includes scores for assignments
that are not yet due (completed early). This makes it more difficult to
identify which grades need to be transferred to the gradebook. This
change omits assignments that are not yet due.