Fix assignment view for an admin user viewing a student's assignments
Currently, the assignment view for an admin user viewing their own
assignments (or any other admin's assignments) shows assignments not
yet assigned. This allows admins to easily see what assignments will
be coming up. However, these extra assignments are currently also shown
when an admin is viewing a student's assignments (as opposed to an
admin's assignments). This change ensures that an admin's view of a
student's assignments will only show the assignments the student sees.