Add feature to allow students to send messages to course staff
This feature allows students to send messages to course staff. Any staff
member can take ownership of a student conversation and reply to it.
Messages and replies viewable on the web app, and e-mailed to both the
student and course staff members who are configured to receive student
messages. Messages and replies are written in a rich text editor, and
may include attachments.

This feature is particularly helpful when there are multiple course
staff members, as it ensures that student messages are sent to all
staff members, and removes the need to coordinate which staff member
will reply to which student message.