Check resource existence after checking for authorization
Currently, the DoesResourceExist controller method runs before checking
for authorization. If DoesResourceExist depends on the status of the
currently signed-in user, it will not work properly if there is no user
currently signed in.

This change ensures that DoesResourceExist runs after checking for
authorization (and potentially after a login redirect, if there is no
currently signed-in user).