Change how e-mail notifications for student messages are sent
Currently, the "from" name on all e-mail notifications is identical
("CS Classroom"). This makes it difficult to identify at a glance who
just replied.

This change causes all e-mail notifications to be sent with the author's
name as the "from" name. Because some e-mail clients ignore distinct
"from" names for the same e-mail address (instead just showing the
first name in the conversation), the "from" e-mail address now varies
as well. This change also includes headers to preserve e-mail threading,
as some e-mail clients will split up conversations if the replier was
not included on the original message.