Properly handle extraneous method overloads
Previously, a submission for a question expecting a single overload of
a method would display a generic error if more than one overload was
specified. This change specifies a more specific error. In addition,
the change combines the logic for handling single- and multi-overload
methods into a single codepath that handles both types of methods.