checkpoint submissions query stuff
Rather than changing the now-shared method GetCheckpointSubmissionsQuery to return a List, go back to it being a Queryable, and see what happens if we avoid doing fancy stuff inside the LINQ query, like a Func that returns true or does a section filter based on the parameter.  This alone may be enough to avoid the extra round trips per user.